// lib/types.ts
export type Socials = {
  instagram?: {
    url?: string;
    handle?: string;
  };
  tiktok?: {
    url?: string;
    handle?: string;
  };
  youtube?: {
    url?: string;
    handle?: string;
  };
  twitch?: {
    url?: string;
    handle?: string;
  };
  x?: {
    url?: string;
    handle?: string;
  };
};